+++ fastchat/utils.py
@@ -1,10 +1,12 @@
+from asyncio import AbstractEventLoop
+import json
 import logging
 import logging.handlers
 import os
-import sys
-import json
-import warnings
 import platform
+import sys
+from typing import AsyncGenerator, Generator
+import warnings
 
 import requests
 import torch
@@ -197,3 +199,29 @@ function() {
     return url_params;
     }
 """
+
+
+def iter_over_async(
+    async_gen: AsyncGenerator, event_loop: AbstractEventLoop
+) -> Generator:
+    """
+    Convert async generator to sync generator
+
+    :param async_gen: the AsyncGenerator to convert
+    :param event_loop: the event loop to run on
+    :returns: Sync generator
+    """
+    ait = async_gen.__aiter__()
+
+    async def get_next():
+        try:
+            obj = await ait.__anext__()
+            return False, obj
+        except StopAsyncIteration:
+            return True, None
+
+    while True:
+        done, obj = event_loop.run_until_complete(get_next())
+        if done:
+            break
+        yield obj
